osx_defaults: Do not make any changes in check mode

+++ lib/ansible/modules/extras/system/osx_defaults.py
@@ -257,6 +257,8 @@ class OSXDefaults(object):
             print "Absent state detected!"
             if self.current_value is None:
                 return False
+            if self.module.check_mode:
+                return True
             self.delete()
             return True
 
@@ -273,6 +275,9 @@ class OSXDefaults(object):
                 return False
         elif self.current_value == self.value:
             return False
+
+        if self.module.check_mode:
+            return True
 
         # Change/Create/Set given key/value for domain in defaults
         self.write()
